refactor(documents): range le pipeline de dépôt sous collectivites/documents
Le pipeline naissait dans `referentiels/preuves/upload/`, où il n'a rien à
faire : il ne connaît ni preuve ni référentiel, et ses trois entrées viennent
des référentiels, des audits et des démarches. Le backend a déjà tranché avec
`collectivites/documents/`, et l'ADR fait de `documents` un sous-domaine de
`collectivites`, aux côtés de `membres` et `personnalisations`.

Ces fichiers naissent dans cette PR : les poser d'emblée au bon endroit ne
déplace rien d'existant, donc ne conflicte avec rien. C'est le déménagement du
dossier `preuves/` qui doit attendre, pas la création d'un fichier neuf.

`useAddFileToLib` suit, renommé au passage en kebab-case : son unique
importateur est le pipeline, et le laisser derrière inversait la dépendance —
un sous-domaine partagé important depuis une feature, le défaut que le
déplacement de `saveBlob` avait déjà corrigé.

`useUploadFileToCollectiviteLibrary` disparaît. Une fois le bucket, les
en-têtes et la déduplication sortis, il ne restait qu'un passe-plat : résoudre
la collectivité courante et dériver l'empreinte. Son unique appelant tient
déjà `collectiviteId`.

Ce retrait rend surtout explicite le choix de la collectivité à chaque dépôt.
L'implicite était un piège : la pièce d'un avis PCAET va dans la bibliothèque
du service instructeur, pas dans celle de la collectivité courante, et un
défaut aurait envoyé le fichier dans une bibliothèque interdite à l'agent.
