+++ apps/app/src/referentiels/preuves/AddPreuveModal/use-file-upload-list.ts
@@ -1,7 +1,7 @@
 import { DocumentHash } from '@tet/domain/collectivites';
 import { useState } from 'react';
 import { FileConstraints, keepWithinMaxFiles } from '../upload/constants';
-import { useUploadFile } from '../upload/use-upload-file';
+import { useUploadFile } from '@/app/collectivites/documents/upload/use-upload-file';
 import { FileUploadItem } from './FileItem';
 import { filesToUploadList, PreparedFile } from './filesToUploadList';
 import { UploadErrorCode, UploadStatus, UploadStatusCode } from './types';
